change: moved tasktile and taskedit to home/helpers
style: started to add style to components
