+++ frontend/src/Components/Home/helpers/TaskEdit/index.jsx
@@ -1,0 +1,9 @@
+import React from 'react';
+
+export default function TaskEdit(props) {
+  return (
+    <div>
+      <h2>Testando</h2>
+    </div>
+  );
+}
