+++ frontend/src/Components/Home/helpers/style/home.css
@@ -1,0 +1,29 @@
+* {
+  font-size: 20px;
+  font-family: sans-serif;
+  padding: 0;
+  margin: 0;
+  box-sizing: border-box;
+}
+
+.task-view {
+  height: 900px;
+}
+
+.task-list {
+  display: flex;
+  flex-direction: row;
+  justify-content: space-between;
+  background-color: beige;
+  flex-wrap: wrap;
+  width: 100%;
+}
+
+.task-list li {
+  list-style-type: none;
+  margin: 30px;
+  width: fit-content;
+  border: 1px solid black;
+  border-radius: 10px;
+  background-color: aliceblue;
+}
